requirements.txt
Faker==40.11.0
numpy==2.4.3
pandas==3.0.1
fastparquet==2025.12.0
pyarrow==23.0.1
matplotlib==3.10.8
seaborn==0.13.2
PyYAML==6.0.3